refactor(lanes): fold pipeline/ into transform/ package; promote stage classes
