--- python/bloqade/lanes/pipeline/__init__.py
@@ -1,4 +1,0 @@
-__all__ = ["LogicalPipeline", "PhysicalPipeline"]
-
-from .logical import LogicalPipeline as LogicalPipeline
-from .physical import PhysicalPipeline as PhysicalPipeline
