--- python/bloqade/lanes/pipeline/physical.py
@@ -1,114 +1,0 @@
-from __future__ import annotations
-
-import warnings
-from dataclasses import dataclass, field
-
-from kirin import passes, rewrite
-from kirin.ir.exception import ValidationErrorGroup
-from kirin.ir.method import Method
-
-from bloqade.gemini.common.validation.terminal_measure import (
-    PhysicalTerminalMeasurementValidation,
-)
-from bloqade.lanes.analysis import layout, placement
-from bloqade.lanes.arch.gemini.physical import get_arch_spec as get_physical_arch_spec
-from bloqade.lanes.arch.spec import ArchSpec
-from bloqade.lanes.heuristics.physical.layout import (
-    PhysicalLayoutHeuristicGraphPartitionCenterOut,
-)
-from bloqade.lanes.heuristics.physical.movement import make_physical_placement_strategy
-from bloqade.lanes.passes import SequentialPlacePass
-from bloqade.lanes.rewrite import circuit2place
-
-from .base import _NativeToPlaceBase, _PlaceToMove
-
-
-@dataclass
-class _PhysicalNativeToPlace(_NativeToPlaceBase):
-    def _post_unroll_validation(self, out: Method, no_raise: bool) -> None:
-        _, errors = PhysicalTerminalMeasurementValidation().run(out)
-        if errors and not no_raise:
-            raise ValidationErrorGroup(
-                f"Physical circuit validation failed with {len(errors)} error(s)",
-                errors=errors,
-            )
-
-    def _lower_qubits(self, out: Method) -> None:
-        rewrite.Walk(circuit2place.RewriteQubitsToPinnedQubits()).rewrite(out.code)
-        rewrite.Walk(circuit2place.RewritePhysicalMeasure()).rewrite(out.code)
-
-
-@dataclass
-class PhysicalPipeline:
-    """Compile a physical squin kernel to the move dialect.
-
-    Qubits are lowered to place.NewPinnedQubit; no logical initialization
-    sequence is inserted.  Validates that the kernel has exactly one terminal
-    measure covering all allocated qubits (post-unroll).
-
-    ``arch_spec`` is the single source of truth: when ``layout_heuristic`` or
-    ``placement_strategy`` are ``None`` (the default), they are constructed in
-    ``emit`` using ``self.arch_spec``, guaranteeing consistency.  Pass explicit
-    instances only when you need a fully custom heuristic or strategy; in that
-    case the caller is responsible for arch-spec consistency.
-    """
-
-    arch_spec: ArchSpec = field(default_factory=get_physical_arch_spec)
-    layout_heuristic: layout.LayoutHeuristicABC | None = None
-    placement_strategy: placement.PlacementStrategyABC | None = None
-    place_opt_type: type[passes.Pass] = field(default=SequentialPlacePass)
-
-    @property
-    def resolved_layout_heuristic(self) -> layout.LayoutHeuristicABC:
-        """Return the active layout heuristic, constructing it from ``arch_spec`` if unset.
-
-        Emits a warning if an explicit heuristic is set whose ``arch_spec``
-        differs from the pipeline's ``arch_spec``.
-        """
-        if self.layout_heuristic is None:
-            return PhysicalLayoutHeuristicGraphPartitionCenterOut(
-                arch_spec=self.arch_spec
-            )
-        if self.layout_heuristic.arch_spec != self.arch_spec:
-            warnings.warn(
-                "PhysicalPipeline.layout_heuristic was constructed with a different "
-                "arch_spec than the pipeline. Initial qubit layout may not match the "
-                "pipeline architecture. Leave layout_heuristic=None to have it built "
-                "automatically from arch_spec.",
-                stacklevel=2,
-            )
-        return self.layout_heuristic
-
-    @property
-    def resolved_placement_strategy(self) -> placement.PlacementStrategyABC:
-        """Return the active placement strategy, constructing it from ``arch_spec`` if unset.
-
-        Emits a warning if an explicit strategy is set whose ``arch_spec``
-        differs from the pipeline's ``arch_spec``.
-        """
-        if self.placement_strategy is None:
-            return make_physical_placement_strategy(arch_spec=self.arch_spec)
-        if self.placement_strategy.arch_spec != self.arch_spec:
-            warnings.warn(
-                "PhysicalPipeline.placement_strategy was constructed with a different "
-                "arch_spec than the pipeline. Compiled moves may not match the pipeline "
-                "architecture. Leave placement_strategy=None to have it built automatically "
-                "from arch_spec, or pass arch_spec= explicitly to "
-                "make_physical_placement_strategy().",
-                stacklevel=2,
-            )
-        return self.placement_strategy
-
-    def emit(self, mt: Method, no_raise: bool = True) -> Method:
-        out = _PhysicalNativeToPlace(arch_spec=self.arch_spec).emit(
-            mt, no_raise=no_raise
-        )
-        self.place_opt_type(out.dialects, no_raise=no_raise)(out)
-
-        out = _PlaceToMove(
-            layout_heuristic=self.resolved_layout_heuristic,
-            placement_strategy=self.resolved_placement_strategy,
-            insert_initialize=False,
-        ).emit(out, no_raise=no_raise)
-
-        return out
